refactor(utils): rename package from util to utils and update package declarations
