+++ src/main/kotlin/lotto/utils/Validator.kt
@@ -1,4 +1,4 @@
-package lotto.util
+package lotto.utils
 
 object Validator {
